Set --docdir to actual destination
CMake 3.19 adds a link to the CMake reference manual in the help
menu, which attempts to open the local file if it exists, and falls
back to the internet if it does not. It searches for this local file
based on CMAKE_DOC_DIR. However, CMAKE_DOC_DIR was not being set to
the actual destination. Change CMAKE_DOC_DIR to reflect the actual
destination so that the CMake GUI can find it.